Fix unable to map port 80 & port 443 with NodePort service type.
Our load balancers currently do not allow us to map a frontend port to a
different backend port number. We therefore need to map port 80 & port 443
directly on the nodes. To be able to do that, we must expand the port range
that the API server allows for node ports.

This should be reverted later. See: https://scm.uninett.no/NIRD/kube/issues/10