Prevent locksmithd from unlocking reboot-lock until kubelet is working
Here we add a unit that must complete before locksmithd can
start. This unit waits for the kube-selftest pod to be running on the
current node before it continues. This ensures that locksmithd will
not unlock the reboot-lock until we know that kubelet is working. That
way we should avoid the cluster entering a broken state during
autoupgrade.