Remove session affinity for API server service.
This session affinity causes problems if one of the API servers is
down. If a client has a connection to the API server that fails, it
will continue to connect to that node, because the session affinity
tries to steer connections back to the failed node.

(There is a related issue that causes a failed API server to never be removed from the list of valid service endpoints. See: https://github.com/kubernetes/kubernetes/issues/22609)