Add some sleeps while applying cluster config.
In order to use certain CRDs we need to wait for them to become
available. This applies to rook and cert-manager, so we apply
these in parts, sleeping in between.

This could be improved later by actually verifying the existence
of the CRD before proceeding instead of sleeping a fixed amount
of seconds.