Uninett changes to Prometheus-operator kube-stage-metrics
- Adjust resource allocation
  Ref: https://github.com/kubernetes/kube-state-metrics#resource-recommendation
- Wire in kube-state-metrics in Prometheus Operator role

Note about tweaking kube-state-metrics resources:

Kube-state-metrics deployment has pod_nanny and addon-resizer as
sidecars. These adjust resources for the pod based on the number of
nodes in the cluster.

To avoid having the deployment changed and pods killed and redeployed
by the resizer every time the manifests for this is applied set the
resources to match the values the resizer use.