Add support for Ansible environments
Trying to follow the structure in https://docs.ansible.com/ansible/latest/user_guide/playbooks_best_practices.html#alternative-directory-layout

Two environments added. One for Safespring and one for a test installation with AWS EKS. Note that the latter is only tested with the prometheus_operator role as teh Terraform parts and a lot of the other ansible roles don't apply to a managed Kubernetes solution.